+++ tests/Parser/Resource/Action/DocumentationTest.php
@@ -78,7 +78,7 @@ class DocumentationTest extends TestCase
 
         foreach ($docs['annotations'] as $name => $data) {
             if (!isset($expected['annotations'][$name])) {
-                $this->fail('A parsed `' . $name . '` annotation was not present in the expected data.');
+                $this->fail('Parsed `' . $name . '` annotations were not present in the expected data.');
             }
 
             foreach ($data as $k => $annotation) {
@@ -86,6 +86,10 @@ class DocumentationTest extends TestCase
                 if ($name === 'param') {
                     // Param annotations are keyed off of the field name.
                     $annotation_key = $annotation['field'];
+                }
+
+                if (!isset($expected['annotations'][$name][$annotation_key])) {
+                    $this->fail('A parsed `' . $name . '` annotation was not present in the expected data.');
                 }
 
                 $this->assertSame(
@@ -165,7 +169,7 @@ class DocumentationTest extends TestCase
                         ]
                     ],
                     'minimum_version' => false,
-                    'responses.length' => 3,
+                    'responses.length' => 5,
                     'annotations' => [
                         'uri' => [
                             [
@@ -236,6 +240,24 @@ class DocumentationTest extends TestCase
                                 'representation' => '\Mill\Examples\Showtimes\Representations\Error',
                                 'version' => false,
                                 'visible' => true
+                            ],
+                            [
+                                'capability' => false,
+                                'description' => 'For no reason.',
+                                'error_code' => false,
+                                'http_code' => '404 Not Found',
+                                'representation' => '\Mill\Examples\Showtimes\Representations\Error',
+                                'version' => '>=1.1.3',
+                                'visible' => true
+                            ],
+                            [
+                                'capability' => false,
+                                'description' => 'For some other reason.',
+                                'error_code' => false,
+                                'http_code' => '404 Not Found',
+                                'representation' => '\Mill\Examples\Showtimes\Representations\Error',
+                                'version' => '>=1.1.3',
+                                'visible' => true
                             ]
                         ]
                     ]
@@ -258,7 +280,7 @@ class DocumentationTest extends TestCase
                         ]
                     ],
                     'minimum_version' => '1.1',
-                    'responses.length' => 5,
+                    'responses.length' => 6,
                     'uri.aliases' => [],
                     'annotations' => [
                         'uri' => [
@@ -435,6 +457,13 @@ class DocumentationTest extends TestCase
                                 'representation' => '\Mill\Examples\Showtimes\Representations\Movie',
                                 'type' => 'object',
                                 'version' => false
+                            ],
+                            [
+                                'description' => false,
+                                'http_code' => '202 Accepted',
+                                'representation' => '\Mill\Examples\Showtimes\Representations\Movie',
+                                'type' => 'accepted',
+                                'version' => '>=1.1.3'
                             ]
                         ],
                         'scope' => [
@@ -496,7 +525,7 @@ class DocumentationTest extends TestCase
                             'version' => false
                         ]
                     ],
-                    'minimum_version' => false,
+                    'minimum_version' => '1.1',
                     'responses.length' => 2,
                     'uri.aliases' => [],
                     'annotations' => [
@@ -508,6 +537,11 @@ class DocumentationTest extends TestCase
                                 'group' => 'Movies',
                                 'path' => '/movies/+id',
                                 'visible' => false
+                            ]
+                        ],
+                        'minVersion' => [
+                            [
+                                'minimum_version' => '1.1'
                             ]
                         ],
                         'uriSegment' => [